requirements.txt
cycler==0.10.0
joblib==0.13.2
jsons==0.8.8

psutil
dill
pprofile
utilitybelt

numpy
scipy
pandas
matplotlib
libnum
kiwisolver==1.1.0
pyparsing==2.4.0
python-dateutil==2.8.0
pytz==2019.1

seaborn==0.9.0
six==1.12.0
tqdm==4.36.1

pynacl==1.5.0
pycryptodomex==3.17.0
